requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiofiles=23.1.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.0.1=pypi_0
anyio=3.7.0=pypi_0
async-timeout=4.0.2=pypi_0
asyncer=0.0.2=pypi_0
attrs=23.1.0=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
coloredlogs=15.0.1=pypi_0
cycler=0.11.0=pypi_0
exceptiongroup=1.1.1=pypi_0
fastapi=0.97.0=pypi_0
ffmpy=0.3.0=pypi_0
filelock=3.12.2=pypi_0
filetype=1.2.0=pypi_0
flatbuffers=23.5.26=pypi_0
fonttools=4.39.4=pypi_0
frozenlist=1.3.3=pypi_0
fsspec=2023.6.0=pypi_0
gradio=3.35.2=pypi_0
gradio-client=0.2.7=pypi_0
h11=0.14.0=pypi_0
httpcore=0.17.2=pypi_0
httpx=0.24.1=pypi_0
huggingface-hub=0.15.1=pypi_0
humanfriendly=10.0=pypi_0
idna=3.4=pypi_0
imagehash=4.3.1=pypi_0
imageio=2.28.1=pypi_0
jinja2=3.1.2=pypi_0
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=pypi_0
lazy-loader=0.2=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
linkify-it-py=2.0.2=pypi_0
llvmlite=0.40.1rc1=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.7.1=pypi_0
mdit-py-plugins=0.3.3=pypi_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.1=pypi_0
numba=0.57.0=pypi_0
numpy=1.24.2=pypi_0
onnxruntime=1.15.1=pypi_0
opencv-python-headless=4.7.0.72=pypi_0
openssl=1.1.1t=h7f8727e_0
orjson=3.9.1=pypi_0
packaging=23.1=pypi_0
pillow=9.5.0=pypi_0
pip=23.0.1=py310h06a4308_0
platformdirs=3.6.0=pypi_0
pooch=1.7.0=pypi_0
protobuf=4.23.3=pypi_0
pydantic=1.10.9=pypi_0
pydub=0.25.1=pypi_0
pygments=2.15.1=pypi_0
pylon=0.4.4=pypi_0
pymatting=1.1.8=pypi_0
pyparsing=3.0.9=pypi_0
pypylon=1.9.0=pypi_0
pyqt5=5.15.9=pypi_0
pyqt5-qt5=5.15.2=pypi_0
pyqt5-sip=12.12.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
readline=8.2=h5eee18b_0
rembg=2.0.43=pypi_0
requests=2.31.0=pypi_0
scikit-image=0.20.0=pypi_0
scipy=1.10.1=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=66.0.0=py310h06a4308_0
six=1.16.0=pypi_0
sniffio=1.3.0=pypi_0
sqlite=3.41.2=h5eee18b_0
starlette=0.27.0=pypi_0
sympy=1.12=pypi_0
tifffile=2023.4.12=pypi_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=pypi_0
tqdm=4.65.0=pypi_0
typing-extensions=4.6.3=pypi_0
tzdata=2023c=h04d1e81_0
uc-micro-py=1.0.2=pypi_0
urllib3=2.0.3=pypi_0
uvicorn=0.22.0=pypi_0
watchdog=3.0.0=pypi_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py310h06a4308_0
xz=5.2.10=h5eee18b_1
yarl=1.9.2=pypi_0
zlib=1.2.13=h5eee18b_0
